+++ javascript/libjoynr-js/src/test/js/joynr/capabilities/TestCapabilitiesRegistrar.js
@@ -145,6 +145,7 @@ joynrTestRequire(
                                 qos : providerQos,
                                 channelId : localChannelId,
                                 participantId : participantId,
+                                publicKeyId: "",
                                 address : address
                             });
                         });